Separate build modules for compiler and web
